Removed `default` argument from interface.
Previousy yaFyaml accessors provided a `default` optional argument
which allows users to provide a default value when the config does not
have a matching entry.  This added a modest amount of complexity and
raised questions about whether the config should then add the entry
for subsquent use.

I decided to instead push that responsibility onto other layers that
use yaFyaml.   E.g., motivated users can write their own wrapper that
allows for default values and potentially insert the new value into
the config if the selector list is not valid.  Or they could simply.

The proper way to use in the case where a selector list might not be
valid is to use the accessors with the `found` optional argument.
Users should check the result of `found` and proceed accordingly.
E.g.,

```f90
call get(config, x, found=found)
if (.not. found) x = my_default
```